Add header computation function

+++ src/cli/cli.odin
@@ -7,8 +7,8 @@ import "core:strings"
 
 Options :: struct {
 	verbose:     bool `args:"name=verbose"                    usage:"Enable verbose output"`,
-	compression: u8 `args:"name=compression"                  usage:"Compression level, 0-12"`,
-	size:        i64 `args:"name=size"                        usage:"Chunk size in bytes"`,
+	compression: u16 `args:"name=compression"                 usage:"Compression level, 0-12"`,
+	size:        u64 `args:"name=size"                        usage:"Chunk size in bytes"`,
 	input:       ^os.File `args:"name=input,file=r,required"  usage:"Path to input directory"`,
 	output:      ^os.File `args:"name=output,file=wc"         usage:"Path to ouput directory"`,
 	// overflow:    [dynamic]string,
